+++ glotaran/io/__init__.py
@@ -34,16 +34,3 @@ from glotaran.plugin_system.project_io_registration import save_scheme
 from glotaran.plugin_system.project_io_registration import set_project_plugin
 from glotaran.plugin_system.project_io_registration import show_project_io_method_help
 from glotaran.utils.io import load_datasets
-
-
-def __getattr__(attribute_name: str):
-    from glotaran.deprecation.deprecation_utils import deprecate_module_attribute
-
-    if attribute_name == "read_data_file":
-        return deprecate_module_attribute(
-            deprecated_qual_name="glotaran.io.read_data_file",
-            new_qual_name="glotaran.io.load_dataset",
-            to_be_removed_in_version="0.6.0",
-        )
-
-    raise AttributeError(f"module {__name__} has no attribute {attribute_name}")